add a switch for Hash in browser/node, revert Base58 to bs58
