+++ lib/encoding/base58.js
@@ -1,8 +1,8 @@
 'use strict'
 
 var _ = require('../util/_')
+var bs58 = require('bs58')
 var buffer = require('buffer')
-var BN = require('../crypto/bn')
 
 /**
  * The alphabet for the Bitcoin-specific Base 58 encoding distinguishes between
@@ -10,8 +10,6 @@ var BN = require('../crypto/bn')
  * prevent accidentaly miscopying of letters.
  */
 var ALPHABET = '123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz'.split('')
-var bn58 = new BN(58)
-var padding = Buffer.from('00', 'hex')
 
 /**
  * A Base58 object can encode/decoded Base 58, which is used primarily for
@@ -57,21 +55,7 @@ Base58.encode = function (buf) {
   if (!buffer.Buffer.isBuffer(buf)) {
     throw new Error('Input should be a buffer')
   }
-  // return bs58.encode(buf)
-  if (buf.length === 0) return ''
-  var leadingZero = 0
-  while (leadingZero < buf.length && buf[leadingZero] === 0) {
-    leadingZero++
-  }
-  var n = BN.fromBuffer(buf)
-  var s = ''
-  var pad = ALPHABET[0].repeat(leadingZero)
-  while (!BN.Zero.eq(n)) {
-    var r = n.divmod(bn58)
-    s = ALPHABET[r.mod.toNumber()] + s
-    n = r.div
-  }
-  return pad + s
+  return bs58.encode(buf)
 }
 
 /**
@@ -84,19 +68,7 @@ Base58.decode = function (str) {
   if (typeof str !== 'string') {
     throw new Error('Input should be a string')
   }
-  // return Buffer.from(bs58.decode(str))
-  if (str === '') return Buffer.from('')
-  var arr = str.split('')
-  var n = BN.Zero
-  var pad = Buffer.from('')
-  while (arr.length > 0) {
-    var c = arr.shift()
-    if (ALPHABET.indexOf(c) === -1) throw new Error('Non-base58 character')
-    var m = new BN(ALPHABET.indexOf(c))
-    if (n.eq(BN.Zero) && m.eq(BN.Zero))pad = Buffer.concat([pad, padding])
-    n = n.mul(bn58).add(m)
-  }
-  return Buffer.concat([pad, n.toBuffer()])
+  return Buffer.from(bs58.decode(str))
 }
 
 Base58.prototype.fromBuffer = function (buf) {
